"Resolved Security Hotspots by adding proper logging" (#145)
* "Resolved Security Hotspots by adding proper logging"

* "Resolved bugs raised by SonarQube"

* "Resolved bugs raised by SonarQube"

+++ src/main/java/org/sunbird/storage/service/StorageServiceImpl.java
@@ -73,16 +73,18 @@ public class StorageServiceImpl implements StorageService {
 	public SBApiResponse uploadFile(MultipartFile mFile, String cloudFolderName) throws IOException {
 		return uploadFile(mFile, cloudFolderName, serverProperties.getCloudContainerName());
 	}
+
 	public SBApiResponse uploadFile(MultipartFile mFile, String cloudFolderName, String containerName) throws IOException {
 		SBApiResponse response = ProjectUtil.createDefaultResponse(Constants.API_FILE_UPLOAD);
 		File file = null;
 		try {
 			file = new File(System.currentTimeMillis() + "_" + mFile.getOriginalFilename());
 			file.createNewFile();
-			FileOutputStream fos = new FileOutputStream(file);
-			fos.write(mFile.getBytes());
-			fos.close();
-			return uploadFile(file, cloudFolderName,containerName);
+			// Use try-with-resources to ensure FileOutputStream is closed
+			try (FileOutputStream fos = new FileOutputStream(file)) {
+				fos.write(mFile.getBytes());
+			}
+			return uploadFile(file, cloudFolderName, containerName);
 		} catch (Exception e) {
 			logger.error("Failed to upload file. Exception: ", e);
 			response.getParams().setStatus(Constants.FAILED);
@@ -90,7 +92,7 @@ public class StorageServiceImpl implements StorageService {
 			response.setResponseCode(HttpStatus.INTERNAL_SERVER_ERROR);
 			return response;
 		} finally {
-			if (file != null) {
+			if (file != null && file.exists()) {
 				file.delete();
 			}
 		}
